[TA4197] fix(replica): preload issue due to large no of extents
Recently we have seen that if volume size is large then preload
takes time and hence openReplica fails due to timeouts and hence
neither of them succedes.

There are five calls where Preload is required
- When replica is in Initial state
- Replica received `open` signal from controller
- Replica is closed
- Replica is reloading
- Revert to snapshot

To solve the issue, we have removed it from the construct function
rather we are calling it after the sync is done already
and successed from controller at the replica side.

NOTE:

Revert was controlled from the controller, that also require preload
to be done. A new PR will be raised to resolve this issue.

Signed-off-by: Utkarsh Mani Tripathi <utkarsh.tripathi@mayadata.io>