[TA4197] fix(replica): preload issue due to large no of extents
Recently we have seen that if volume size is large then preload
takes time and hence openReplica fails due to timeouts and hence
neither of them succedes.

There are five calls where Preload is required

* When replica is in Initial state
* Replica received `open` signal from controller
* Replica is closed
* Replica is reloading
* Revert to snapshot

To solve the issue, we have removed http client timeout, and fixed
other issues which was depending on this.

Feature : Added `PreloadStatus` field in Replica structure, so status can be queried via REST api
There are three status:

```
Error    : In case of replica disconnection or syscall failure
Started  : If Preload is happening in the background
Done     : If Preload is completed.
```

Signed-off-by: Utkarsh Mani Tripathi <utkarsh.tripathi@mayadata.io>