fix(replica): bad file descriptor error
Recently it has been observed that replicas were crashing due to
frequent controller restart. This is happening due to the files
are getting closed just after the disconnection but it is trying to
punch holes on the closed files. So it was giving `bad file descriptor
error`.

This PR fix the above issue by closing the files gracefully after
draining the queued data in the HoleCreatorChan. Punch holes will
be done only in following case:
- If replica is master replica (which get start signal)
- If replica is reloaded
- If replica is in RW mode (i.e, Rebuilding is false)

Additionally it fixes some go vet errors also.

Signed-off-by: Utkarsh Mani Tripathi <utkarsh.tripathi@mayadata.io>