fix(replica): fix broken fs due to delete snapshot
Fix: openebs/jiva#230

This commit fixes the issue of broken file system at the
application side.

This is happening due to the hardlinked files were pointing
to the older fd's which is not valid anymore after the hardlinking
of child to parent snapshot. So files need to be reopened and there
reference should be updated with the new one.

Signed-off-by: Utkarsh Mani Tripathi <utkarsh.tripathi@mayadata.io>