feat(replica): run fsync on files & dir after create/remove/rename operation on files (#278)
We open file in O_DIRECT mode which minimizes the cache effects but doesn't
guarantee that data is written to the disk. Periodic sync ensures that both
data and metadata is flushed to the disk at every 5 seconds. See man page of
open system call for more details. Also since fsync on fd's also doesn't
guarantee the same so fsync on directory is also required.

In this commit a goroutine is launched to perform sync on data files and
directory. Also sync and close the replicas gracefully before fataling upon
error or killing the replicas.

Signed-off-by: Utkarsh Mani Tripathi <utkarsh.tripathi@mayadata.io>