refactor(build): trim leading v from image tag
Using the github tag, docker image tags are
created. As we move the github tags to semver
format, the image tags will end up having a
leading `v` in the tag.

To keep the image tag names consistent with
past releases, the leading `v` will be trimmed
from the travis tag and the rest of the string
will be used for image tag.

Examples:
1.10.0 maps to 1.10.0
v1.10.0 maps to 1.10.0
v1.10.0-custom-RC1 maps to 1.10.0-custom-RC1

Signed-off-by: kmova <kiran.mova@mayadata.io>